Reduce duplication ratio
