+++ source/Handlebars/Compiler/Translation/Expression/HelperFunctionBinder.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Linq.Expressions;
-using System.Reflection;
 using System.Runtime.CompilerServices;
 using Expressions.Shortcuts;
-using HandlebarsDotNet.Collections;
 using HandlebarsDotNet.Helpers;
 using static Expressions.Shortcuts.ExpressionShortcuts;
 
@@ -13,8 +8,6 @@ namespace HandlebarsDotNet.Compiler
 {
     internal class HelperFunctionBinder : HandlebarsExpressionVisitor
     {
-        private static readonly LookupSlim<int, DeferredValue<Expression[], ConstructorInfo>> ArgumentsConstructorsMap = new LookupSlim<int, DeferredValue<Expression[], ConstructorInfo>>();
-        
         private CompilationContext CompilationContext { get; }
 
         public HelperFunctionBinder(CompilationContext compilationContext)
@@ -37,7 +30,7 @@ namespace HandlebarsDotNet.Compiler
             var textWriter = Arg<EncodedTextWriter>(CompilationContext.EncodedWriter);
             
             var contextValue = bindingContext.Property(o => o.Value);
-            var args = CreateArguments();
+            var args = FunctionBinderHelpers.CreateArguments(hex.Arguments, CompilationContext);
 
             var configuration = CompilationContext.Configuration;
             if (configuration.Helpers.TryGetValue(pathInfo, out var helper))
@@ -60,42 +53,6 @@ namespace HandlebarsDotNet.Compiler
             configuration.Helpers.Add(pathInfo, lateBindDescriptor);
             
             return Call(() => lateBindDescriptor.Value.WriteInvoke(bindingContext, textWriter, contextValue, args));
-            
-            ExpressionContainer<Arguments> CreateArguments()
-            {
-                Expression[] arguments = hex.Arguments
-                    .ApplyOn<Expression, PathExpression>(path => path.Context = PathExpression.ResolutionContext.Parameter)
-                    .Select(o => FunctionBuilder.Reduce(o, CompilationContext))
-                    .ToArray();
-
-                if (arguments.Length == 0)
-                {
-                    return Arg(Arguments.Empty);
-                }
-
-                var constructor = ArgumentsConstructorsMap.GetOrAdd(arguments.Length, (i, d) =>
-                {
-                    return new DeferredValue<Expression[], ConstructorInfo>(d, o =>
-                    {
-                        var objectType = typeof(object);
-                        var argumentTypes = new Type[o.Length];
-                        for (var index = 0; index < argumentTypes.Length; index++)
-                        {
-                            argumentTypes[index] = objectType;
-                        }
-
-                        return typeof(Arguments).GetConstructor(argumentTypes);
-                    });
-                }, arguments).Value;
-
-                if (constructor == null) // cannot handle by direct args pass
-                {
-                    var arr = Array<object>(arguments);
-                    return New(() => new Arguments(arr));
-                }
-                
-                return Arg<Arguments>(Expression.New(constructor, arguments));
-            }
         }
     }
 }
